+++ application/controllers/backoffice/CriteriaManagement.php
@@ -70,70 +70,73 @@ class CriteriaManagement extends AdminController
      */
     public function addEditCriteria()
     {
+        //check entry in entry_record
+        $entry_record = $this->CommonModel->getRecord('entry_record')->num_rows();
+        
+        if($entry_record == 0) {
+            if ($this->input->post('action') && $this->input->post('action') == "addCriteria") {
+                $criteria_data = array(
+                    "section_id" => $this->input->post('criteria_frm_section_id'),
+                    "criteria_name" => $this->input->post('criteria_frm_criteria_name')
+                );
+                if ($this->input->post('radios') == "options") {
+                    $criteria_data['type_data'] = 1;
+                }
 
-        if ($this->input->post('action') && $this->input->post('action') == "addCriteria")
-        {
-            $criteria_data = array(
-                "section_id" => $this->input->post('criteria_frm_section_id'),
-                "criteria_name" => $this->input->post('criteria_frm_criteria_name')
-            );
-            if($this->input->post('radios') == "options"){
-                $criteria_data['type_data'] = 1;
+
+                $save = $this->CommonModel->save("criteria_master", $criteria_data);
+
+
+                if ($save) {
+
+                    //entry in option_master if options are available
+                    if ($this->input->post('radios') == "options") {
+                        $options = $this->input->post('options');
+                        foreach ($options as $row):
+                            $row['criteria_id'] = $save;
+                            $option_data = $row;
+                            unset($option_data['section_id']);
+                            $this->CommonModel->save("option_master", $option_data);
+                        endforeach;
+                    }
+                    $this->session->set_flashdata("success", "Criteria added successfully");
+                } else {
+                    $this->session->set_flashdata("error", "problem adding Criteria. Try Later");
+                }
             }
 
-
-
-            $save = $this->CommonModel->save("criteria_master",$criteria_data);
-
-
-            if($save){
-
+            if ($this->input->post('action') && $this->input->post('action') == "editCriteria") {
+                $criteria_data = array(
+                    "section_id" => $this->input->post('criteria_frm_section_id'),
+                    "criteria_name" => $this->input->post('criteria_frm_criteria_name'),
+                    "type_data" => 0
+                );
+                if ($this->input->post('radios') == "options") {
+                    $criteria_data['type_data'] = 1;
+                }
                 //entry in option_master if options are available
-                if($this->input->post('radios') == "options"){
+                if ($this->input->post('radios') == "options") {
+                    //delete previous if any
+                    $this->CommonModel->delete("option_master", array('criteria_id' => $this->input->post('update_id')));
                     $options = $this->input->post('options');
                     foreach ($options as $row):
-                        $row['criteria_id'] = $save;
+                        $row['criteria_id'] = $this->input->post('update_id');
                         $option_data = $row;
                         unset($option_data['section_id']);
-                        $this->CommonModel->save("option_master",$option_data);
+                        $this->CommonModel->save("option_master", $option_data);
                     endforeach;
                 }
-                $this->session->set_flashdata("success","Criteria added successfully");
-            }else{
-                $this->session->set_flashdata("error","problem adding Criteria. Try Later");
+                $update = $this->CommonModel->update("criteria_master", $criteria_data, array('criteria_id' => $this->input->post('update_id')));
+                if ($update) {
+                    $this->session->set_flashdata("success", "Criteria updated successfully");
+                } else {
+                    $this->session->set_flashdata("error", "Problem Editing Criteria.Try Later");
+                }
             }
         }
-        
-        if ($this->input->post('action') && $this->input->post('action') == "editCriteria")
-        {
-            $criteria_data = array(
-                "section_id" => $this->input->post('criteria_frm_section_id'),
-                "criteria_name" => $this->input->post('criteria_frm_criteria_name'),
-                "type_data" => 0
-            );
-            if($this->input->post('radios') == "options"){
-                $criteria_data['type_data'] = 1;
-            }
-            //entry in option_master if options are available
-            if($this->input->post('radios') == "options"){
-                //delete previous if any
-                $this->CommonModel->delete("option_master",array('criteria_id'=>$this->input->post('update_id')));
-                $options = $this->input->post('options');
-                foreach ($options as $row):
-                    $row['criteria_id'] = $this->input->post('update_id');
-                    $option_data = $row;
-                    unset($option_data['section_id']);
-                    $this->CommonModel->save("option_master",$option_data);
-                endforeach;
-            }
-            $update = $this->CommonModel->update("criteria_master",$criteria_data,array('criteria_id'=>$this->input->post('update_id')));
-            if($update){
-                $this->session->set_flashdata("success","Criteria updated successfully");
-            }else{
-                $this->session->set_flashdata("error","Problem Editing Criteria.Try Later");
-            }
+        else{
+            $this->session->set_flashdata("error", "Entry Record Exist . First Delete That Records");
         }
-        
         redirect("backoffice/CriteriaManagement","refresh");
     }
     
@@ -145,6 +148,7 @@ class CriteriaManagement extends AdminController
      */
     public function viewEditCriteriaModal($criteria_id)
     {
+
         $OrWhere = array();
         $section_list = $this->CommonModel
             ->dbOrderBy(array('section_id'=>'ASC'))
@@ -176,20 +180,31 @@ class CriteriaManagement extends AdminController
     {
         if ($this->input->post('criteria_id'))
         {
-            $result = $this->CommonModel->delete("criteria_master",array('criteria_id'=>$this->input->post('criteria_id')));
-            if ($result)
-            {
-                $res_output['code'] = 1;
-                $res_output['status'] = "success";
-                $res_output['message'] = "Criteria deleted successfully";
-                echo json_encode($res_output);
-                exit();
-            }
-            else 
-            {
+            //check entry in entry_record
+            $entry_record = $this->CommonModel->getRecord('entry_record')->num_rows();
+            if($entry_record == 0){
+                $result = $this->CommonModel->delete("criteria_master",array('criteria_id'=>$this->input->post('criteria_id')));
+                if ($result)
+                {
+                    $res_output['code'] = 1;
+                    $res_output['status'] = "success";
+                    $res_output['message'] = "Criteria deleted successfully";
+                    echo json_encode($res_output);
+                    exit();
+                }
+                else
+                {
+                    $res_output['code'] = 0;
+                    $res_output['status'] = "error";
+                    $res_output['message'] = "Criteria not delete";
+                    echo json_encode($res_output);
+                    exit();
+                }
+            }else{
+                //record exist in entry record
                 $res_output['code'] = 0;
                 $res_output['status'] = "error";
-                $res_output['message'] = "Criteria not delete";
+                $res_output['message'] = "Entry Record Exist . First Delete That Records";
                 echo json_encode($res_output);
                 exit();
             }
